added more configuration to button configuration
