+++ Sources/UIComponents/Components/Buttons/PrimaryButton.swift
@@ -10,6 +10,7 @@ extension GOVUKButton.ButtonConfiguration {
     public static var primary: GOVUKButton.ButtonConfiguration {
         .init(titleNormal: .white,
               titleFocused: .black,
+              textAlignment: .left,
               backgroundNormal: UIColor(resource: .myBlue),
               backgroundFocused: .systemYellow,
               backgroundShape: .capsule)
